2 more

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -7,9 +7,9 @@
  * Return: 0 success.
  */
 
-int mul(long a, long b)
+int mul(int a, int b)
 {
-	long mult, divisor, digit;
+	int mult, divisor, digit;
 
 	mult = a * b;
 	divisor = 1;
